+++ src/ContaoManager/Plugin.php
@@ -18,7 +18,7 @@ use Contao\ManagerPlugin\Bundle\Config\BundleConfig;
 use Contao\ManagerPlugin\Bundle\Config\ConfigInterface;
 use Contao\ManagerPlugin\Bundle\Parser\ParserInterface;
 use Contao\ManagerPlugin\Config\ConfigPluginInterface;
-use HeimrichHannot\SearchBundle\ContaoSearchBundle;
+use HeimrichHannot\SearchBundle\HeimrichHannotSearchBundle;
 use Symfony\Component\Config\Loader\LoaderInterface;
 
 class Plugin implements BundlePluginInterface
@@ -32,7 +32,7 @@ class Plugin implements BundlePluginInterface
     public function getBundles(ParserInterface $parser)
     {
         return [
-          BundleConfig::create(ContaoSearchBundle::class)->setLoadAfter([
+          BundleConfig::create(HeimrichHannotSearchBundle::class)->setLoadAfter([
              ContaoCoreBundle::class
           ])
         ];