+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 
 ## [DEV] - 2020-02-06
 * added RebuildSearchIndexCommand
+* added disable search index option
+* [BC BREAK] Renamed bundle class
 
 ## [1.1.2] - 2020-01-08
 * adapted to coding standards